Test snmp listener tags
